Store Orders: Add new components for the status label & select dropdown (#16930)
* Store: Create components for order status- one to display a label, the other to display a dropdown of available order statuses

* Fix an error trying to use indexOf on an undefined value

* Add README
