+++ client/extensions/woocommerce/components/order-status/style.scss
@@ -1,0 +1,42 @@
+.order-status__item {
+	display: inline-flex;
+	padding: 0px 12px;
+	line-height: 32px;
+	background: lighten( $gray, 20% );
+	border-radius: 4px;
+
+	&.is-failed {
+		background: lighten( $alert-red, 20% );
+		color: darken( $alert-red, 30% );
+
+		span + span {
+			border-color: $alert-red;
+		}
+	}
+
+	&.is-cancelled,
+	&.is-refunded,
+	&.is-completed {
+		background: $gray-light;
+		color: $gray-dark;
+
+		span + span {
+			border-color: lighten( $gray, 20% );
+		}
+	}
+
+	&.is-on-hold {
+		background: lighten( $alert-yellow, 20% );
+		color: darken( $alert-yellow, 30% );
+
+		span + span {
+			border-color: $alert-yellow;
+		}
+	}
+
+	span + span {
+		margin-left: 12px;
+		padding-left: 12px;
+		border-left: 1px solid $gray;
+	}
+}
