+++ client/extensions/woocommerce/style.scss
@@ -22,6 +22,7 @@
 	@import 'app/dashboard/style';
 	@import 'app/order/style';
 	@import 'app/orders/style';
+	@import 'components/order-status/style';
 	@import 'components/basic-widget/style';
 	@import 'components/share-widget/style';
 	@import 'components/reading-widget/style';
